+++ src/piddiplatsch/core/processing.py
@@ -14,11 +14,13 @@ class BaseProcessor:
         dry_run: bool = False,
         project: str | None = None,
         handle_profile: str | None = None,
+        handle_output_filename: str | None = None,
     ):
         self.handle_backend: HandleAPIProtocol = handle_backend or HandleAPI(
             dry_run=dry_run,
             project=project,
             handle_profile=handle_profile,
+            output_filename=handle_output_filename,
         )
         self.handle_profile = handle_profile
         self.retries = retries